# contracts/quality-api/openapi.yaml
openapi: 3.1.0
...
info:
  title: CaseLoop Quality API
  version: "2.0.0"
  summary: AI 应用质量自治底座——被治理应用的质量契约（v2 强化版）
  description: |
    Quality API v2（plan-v3 §2.3.1 / T0）。任何 LLM 应用实现本契约即可被 CaseLoop 纳管。

    核心不变量：
    - **VersionSet 不可变**：{prompt, KB manifest, model+params} 创建后内容不可改，
      变更 = 新 VersionSet（新 id + 新 digest）；只有生命周期状态可迁移。
    - **生命周期**：draft → stage → canary → promote | rollback → status。
    - **写面纪律**：全部写操作要求 CAS（If-Match 头 或 body expected_revision）、
      Idempotency-Key 头，且为异步——返回 operation_id，经 GET /v2/operations/{id} 查结果。
    - **写面仅 Release Controller 可调**：quality:write scope 令牌只签发给
      Release Controller（控制面组件，非 LLM）。应用自身、Agent 均不得持有。
    - **读面**：GET /v2/logs、GET /v2/feedback（quality:read 或 quality:write）。

    digest 规则：所有 digest = sha256(JCS(RFC 8785) canonical JSON)，
    表示为 `sha256:<64 hex>`；VersionSet 的完整版本 digest 覆盖其
    {prompt, kb_manifest, model} 三元组内容。

  contact:
    name: CaseLoop contracts
    url: https://caseloop.dev/contracts
servers:
  - url: http://127.0.0.1:8080
    description: demo-app 本地默认（conformance suite 的 BASE_URL 缺省值）
...
paths:
...
  /v2/operations/{operationId}:
    get:
      tags: [operations]
      summary: 查询异步写操作结果
      description: |
        写操作（stage/canary/promote/rollback）均为异步：先返 202 + operation_id，
        本端点轮询至 succeeded/failed。operation 记录有 TTL（【待定】默认 24h），
        过期清理后查询返回 410 operation_expired——发布侧 UNKNOWN 语义见此：
        调用方查不到终态时必须按 Release 状态机走 reconcile，不得盲目重试。
      operationId: getOperation
      security:
        - quality_oauth: [quality:read]
        - quality_oauth: [quality:write]
      parameters:
        - name: operationId
          in: path
          required: true
          schema:
            type: string
            pattern: "^op_[0-9A-Za-z]{8,64}$"
      responses:
        "200":
          description: 操作当前状态
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Operation"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "410":
          $ref: "#/components/responses/OperationExpired"
...
components:
  securitySchemes:
    quality_oauth:
      type: oauth2
      description: |
        OAuth2 client credentials。scope 语义：
        - **quality:read**：读面（GET logs/feedback/versionsets/operations）。
        - **quality:write**：写面（创建 VersionSet、全部生命周期迁移、admin 注入）。
          **写 scope 令牌仅签发给 Release Controller**（plan-v3 §2.3.1 硬约束）；
          持有 quality:write 的调用方同时视为有读权限。
      flows:
        clientCredentials:
          tokenUrl: /oauth/token
          scopes:
            quality:read: "读面：logs/feedback/versionsets/operations 查询"
            quality:write: "写面：仅 Release Controller 可持有"
...
  responses:
...
    Unauthorized:
      description: 未认证（令牌缺失/无效）
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            error: { code: unauthorized, message: "missing or invalid token" }
...
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            error: { code: not_found, message: "versionset not found" }
...
    OperationExpired:
      description: 410 operation_expired——operation 记录已过 TTL 被清理，终态不可考（调用方按 UNKNOWN→reconcile 处理）
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            error:
              code: operation_expired
              message: "operation record expired and purged"
  schemas:
    Digest:
      type: string
      description: sha256(JCS canonical JSON) 内容指纹
      pattern: "^sha256:[0-9a-f]{64}$"
      example: "sha256:1e79b1bdcf3d5a931980fa28a7a4c0313d7aecf0f4d1de08df4d721927da8936"
...
    LifecycleStatus:
      type: string
      description: |
        生命周期状态机（非法迁移一律 422 validation_failed）：
        draft → staged → canary → active（promote）；canary/active → rolled_back（rollback）；
        promote 时原 active 自动 → superseded。
      enum: [draft, staged, canary, active, superseded, rolled_back]
...
    Operation:
      type: object
      additionalProperties: false
      required: [operation_id, kind, status, idempotency_key, created_at, expires_at]
      properties:
        operation_id:
          type: string
          pattern: "^op_[0-9A-Za-z]{8,64}$"
        kind:
          type: string
          enum: [stage, canary, promote, rollback]
        status:
          type: string
          description: pending/running 为中间态；succeeded/failed 为终态
          enum: [pending, running, succeeded, failed]
        idempotency_key:
          type: string
        versionset_id:
          type: string
        result:
          type: object
          description: 成功终态的结果
          additionalProperties: false
          properties:
            revision:
              type: integer
            status:
              $ref: "#/components/schemas/LifecycleStatus"
            restored_digest:
              allOf:
                - $ref: "#/components/schemas/Digest"
              description: "rollback 成功时必填；实际恢复为 active 的 VersionSet digest"
        error:
          $ref: "#/components/schemas/ErrorBody"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
          description: 记录 TTL（【待定】默认 24h），过后查询返回 410 operation_expired
...
    ErrorBody:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
          description: |
            机器可判错误码。HTTP 映射：400→bad_request；401→unauthorized；403→forbidden；
            404→not_found；409→revision_conflict；410→operation_expired；
            412→precondition_failed；422→validation_failed。
          enum:
            - bad_request
            - unauthorized
            - forbidden
            - not_found
            - revision_conflict
            - precondition_failed
            - validation_failed
            - operation_expired
        message:
          type: string
          description: 人读描述
        details:
          type: object
          description: 结构化补充（subcode / expected_revision / current_revision / current_status / attempted 等）
          additionalProperties: true
        operation_id:
          type: string
        trace_id:
          type: string
    Error:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/ErrorBody"
security: []